refactor(game): split 1036-line tiles map page into focused files under 250 lines
Reduces app/game/tiles/page.tsx from 1036 lines to 201 by extracting:

- 8 presentation components into app/game/tiles/_components/
  (TileHexagon — single hex render with all decorations,
  TileHoverCard, ZoomControls, MapCanvas — the SVG + zoom controls + hover,
  PersonalMapToolbar, ScopeFilterRow, LandTypeFilterRow, MapLegend)

- 5 lib modules into app/game/tiles/_lib/
  (use-tiles-data hook — player/cachedView/worldView/mode state,
  use-pan-zoom hook — scale/tx/ty + wheel + drag,
  hex-math — pure axial-to-pixel + viewport-fit math,
  types — response shapes, ScopeFilter, ViewMode,
  constants — colors, sizes, filter lists)

The _components and _lib prefixes are Next.js private folders, so the new
structure does not introduce any new routes. page.tsx is now a thin
orchestrator that binds the data hook + pan-zoom hook to the composed
MapCanvas and handles only the modal-tile state.

Adds unit tests for the pure hex-math helpers (axialToPixel, hexPoints,
fitTilesToViewport) including the "never zoom past 1:1 on recenter"
invariant.

No behavior change. Tests pass; tsc clean; lint clean (no new warnings).
